Added some MoreMusic functions and checks
The checks are far from finished.  I'm still trying to work out how to write shrinkers properly.  At the moment, Producers for PAR and SEQ in Music Pitch crash because of the recursion (with elm not being lazy).  I don't yet know how to circumvent this.

Also slightly unhappy about using imeckler/Ratio because the constructor is hidden and so you cannot pattern match on Rationals.  I may need to fork this library.