Upgraded to latest version of elm-test
This deprecates elm-check and replaces the old idea of a Producer with the new one of a Fuzzer.  All QuickCheck style checks can be placed inside a conventional elm-test suite.  The tests can be run simply by changing to the home directory and typing elm-test.  Note that the latest elm-test runner must first be installed - see documentation in http://package.elm-lang.org/packages/elm-community/elm-test/2.0.1/.