Upgrade to Elm 0.18 (#9)
* Added some MoreMusic functions and checks

The checks are far from finished.  I'm still trying to work out how to write shrinkers properly.  At the moment, Producers for PAR and SEQ in Music Pitch crash because of the recursion (with elm not being lazy).  I don't yet know how to circumvent this.

Also slightly unhappy about using imeckler/Ratio because the constructor is hidden and so you cannot pattern match on Rationals.  I may need to fork this library.

* Made checks slightly dryer

* Upgraded to latest version of elm-test

This deprecates elm-check and replaces the old idea of a Producer with the new one of a Fuzzer.  All QuickCheck style checks can be placed inside a conventional elm-test suite.  The tests can be run simply by changing to the home directory and typing elm-test.  Note that the latest elm-test runner must first be installed - see documentation in http://package.elm-lang.org/packages/elm-community/elm-test/2.0.1/.

* Added some MEvent functions

This has depended on a forked version of imeckler/Ratio which supports mixed-type arithmetic.

* Ported phraseToMEvents

(which is a monster)

Also added a few more to Ratio to make life slightly more pleasant.

* Almost completed MEvent

Added a pseudo typeclass Music1 (probably not much use)
Upgrade to latest newlandsvalley/Ratio

* Upgrade to elm 0.18

* Revert test-runner package to 3.0.0

(rtfeldman had corrected it)

* Replace locat ratio with elm-community/ratio

* Amended README